Add feature for using AsmMachine in debugger
The consensus is now defined as the x64 assembly implementation of
ckb-vm, having an alternative debugger binary that uses the assembly VM,
can IMHO be helpful in certain cases. We all know the Rust version is
easier to plug-in all kinds of utilities, but the assembly version will
always have a handy place avoiding confusions.